[libromdata] Wim: Cast rowloop_current_windowsinfo.arch to int for rp_sprintf().
This only showed up if -DENABLE_NLS=OFF:

src/libromdata/Other/Wim.cpp: In member function ‘int LibRomData::WimPrivate::addFields_XML()’:
src/libromdata/Other/Wim.cpp:369:72: error: format ‘%u’ expects argument of type ‘unsigned int’, but argument 2 has type ‘int’ [-Werror=format=]
  369 |                         data_row.emplace_back(rp_sprintf(C_("RomData", "Unknown (%u)"), rowloop_current_windowsinfo.arch));
src/libromdata/../libi18n/i18n.h:34:50: note: in definition of macro ‘C_’
   34 | #  define C_(msgctxt, msgid)                    (msgid)
      |                                                  ^~~~~
cc1plus: some warnings being treated as errors

+++ src/libromdata/Other/Wim.cpp
@@ -366,7 +366,8 @@ int WimPrivate::addFields_XML()
 		if (archstring) {
 			data_row.emplace_back(archstring);
 		} else {
-			data_row.emplace_back(rp_sprintf(C_("RomData", "Unknown (%u)"), rowloop_current_windowsinfo.arch));
+			data_row.emplace_back(rp_sprintf(C_("RomData", "Unknown (%d)"),
+				static_cast<int>(rowloop_current_windowsinfo.arch)));
 		}
 		data_row.emplace_back(rowloop_current_windowsinfo.languages.language);
 	}
